extract nodes and links data model in separated modules
